app/requirements.txt
aiogram-3.5.0
Django==4.2.11
django-jet-reboot==1.3.7
djangorestframework==3.15.1
pillow==10.3.0